requirements.txt
numpy
plotly-express
pandas
autopep8
scikit-learn
keras
tensorflow